fix(server): 종료된 송신자의 남은 세션 복구

+++ src/server.c
@@ -1,33 +1,59 @@
 #include "minitalk.h"
 
+#include <errno.h>
 #include <unistd.h>
 
 static volatile sig_atomic_t	g_current_byte;
 static volatile sig_atomic_t	g_received_bits;
 static volatile sig_atomic_t	g_client_pid;
+static volatile sig_atomic_t	g_line_started;
+
+static void	reset_session(int close_partial_line)
+{
+	if (close_partial_line && g_line_started)
+		write(STDOUT_FILENO, "\n", 1);
+	g_current_byte = 0;
+	g_received_bits = 0;
+	g_client_pid = 0;
+	g_line_started = 0;
+}
 
 static void	flush_byte(unsigned char output)
 {
 	if (output == '\0')
 	{
 		write(STDOUT_FILENO, "\n", 1);
-		g_client_pid = 0;
+		reset_session(0);
 	}
 	else
+	{
 		write(STDOUT_FILENO, &output, 1);
+		g_line_started = 1;
+	}
 }
 
 static void	handle_bit(int signal, siginfo_t *info, void *context)
 {
 	unsigned char	output;
+	int				saved_errno;
 
+	saved_errno = errno;
 	(void)context;
 	if (info == NULL || info->si_pid <= 0)
+	{
+		errno = saved_errno;
 		return ;
+	}
 	if (g_client_pid != 0 && g_client_pid != info->si_pid)
 	{
-		kill(info->si_pid, MT_NACK_SIGNAL);
-		return ;
+		if (kill((pid_t)g_client_pid, 0) == -1 && errno == ESRCH)
+			reset_session(1);
+		else
+		{
+			kill(info->si_pid, MT_NACK_SIGNAL);
+			errno = saved_errno;
+			return ;
+		}
 	}
 	if (g_client_pid == 0)
 		g_client_pid = info->si_pid;
@@ -42,7 +68,9 @@ static void	handle_bit(int signal, siginfo_t *info, void *context)
 		g_current_byte = 0;
 		g_received_bits = 0;
 	}
-	kill(info->si_pid, MT_ACK_SIGNAL);
+	if (kill(info->si_pid, MT_ACK_SIGNAL) == -1 && errno == ESRCH)
+		reset_session(1);
+	errno = saved_errno;
 }
 
 static int	install_signal_handlers(void)
